[modmail] extracted tagging to own class
